Create create_mac_file.py
This script will take those mac address, and write them to a file. Which can later be called to compare the OUI to the first 6 charters that are being learned on a switch port. The list 'dell' also contains mac address from Microsoft.

This is useful for when you need to see which ports on a switch contain Dell or Microsoft PCs on them.